Minor refactoring, move impls from `command` to `cli` where structs are defined
